refactor(server): extract routers/reports.py + heavy_tests.py + assessments.py (issue #358)
Move 8 routes from server.py into three new router files:
- routers/reports.py: GET /api/reports, /{date}, /{date}/chart
- routers/heavy_tests.py: GET /api/heavy-tests/repos, POST /dispatch, POST /docker
- routers/assessments.py: GET /api/assessments/scores, POST /dispatch

Dependencies (run_cmd, REPORTS_DIR, HEAVY_TEST_REPOS) injected at startup.
server.py line count drops by ~250 lines.

Closes #358
